Increase modularity of the exchange_processor.
For the methods `yield_exchange_price_updates` and
`yield_exchange_rsi_updates` created a utility `_process_exchanges`
method that takes in the processed_exchanges, exchanges to process, the
method used to process them, and the method used to generate an output
for the results of the processing.

Both `yield` methods used essentially the same logic, just the methods
used to process the data was different so I passed them in, reducing
code duplication.